requirements.txt
uvicorn==0.34.0
Flask==3.0.0
Flask-Cors==5.0.0
FastAPI==0.115.6
Rpi.GPIO
w1thermsensor
werkzeug>=3.0.6 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability